requirements.txt
SAMITorch>=0.1.15
pytorch-ignite>=0.2.0
PyYAML>=5.1
pytorch-ignite>=0.2.0